gem-healthcare-pro/requirements.txt
google-generative-ai
google-generative-agent
google-cloud-firestore
google-cloud-logging


